Descriptive json and move to util.js
