+++ src/endpoints/image-metadata.js
@@ -12,7 +12,7 @@ import writeFileAtomic from 'write-file-atomic';
 import { Jimp } from '../jimp.js';
 import { getConfigValue } from '../util.js';
 
-export const METADATA_FILE = 'index.json';
+export const METADATA_FILE = 'image-metadata.json';
 
 /**
  * @typedef {Object} ImageMetadata
